requirements.txt
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.6.0
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
boto3==1.26.137
botocore==1.29.137
certifi==2024.7.4
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
cryptography==40.0.2
dataclasses-json==0.6.7
deepdiff==7.0.1
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==4.0.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
emoji==2.12.1
filelock==3.15.4
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
groq==0.9.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.4
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-groq==0.1.9
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.99
lxml==5.3.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
ordered-set==4.1.0
orjson==3.10.7
packaging==24.1
pgvector==0.3.2
pillow==10.4.0
psutil==6.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.7.0
pypdf==4.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
python-iso639==2024.4.27
python-magic==0.4.27
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.2
rapidfuzz==3.9.6
regex==2024.7.24
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
s3transfer==0.6.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
sendgrid==6.11.0
sentence-transformers==3.0.1
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.6
SQLAlchemy==2.0.32
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.15.5
unstructured-client==0.25.5
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
